+++ terrasnek/policies.py
@@ -35,8 +35,10 @@ class TFCPolicies(TFCEndpoint):
         """
         ``GET /organizations/:organization_name/policies``
 
-        PARAMS:
-            https://www.terraform.io/docs/cloud/api/policies.html#query-parameters
+        Query parameter(s) (`details <https://www.terraform.io/docs/cloud/api/policies.html#query-parameters>`_):
+            - ``page`` (Optional)
+            - ``page_size`` (Optional)
+            - ``search`` (Optional)
         """
         return self._list(\
             self._org_api_v2_base_url, page=page, page_size=page_size, search=search)