+++ terrasnek/policy_sets.py
@@ -40,8 +40,23 @@ class TFCPolicySets(TFCEndpoint):
         """
         ``GET /organizations/:organization_name/policy-sets``
 
-        PARAMS:
-            https://www.terraform.io/docs/cloud/api/policy-sets.html#list-policy-sets
+        Query parameter(s) (`details <https://www.terraform.io/docs/cloud/api/policy-sets.html#query-parameters>`_):
+            - ``filter[versioned]`` (Optional)
+            - ``include`` (Optional)
+            - ``page`` (Optional)
+            - ``page_size`` (Optional)
+            - ``search`` (Optional)
+
+        Example filter(s):
+
+        .. code-block:: python
+
+            filters = [
+                {
+                    "keys": ["versioned"],
+                    "value": "foo"
+                }
+            ]
         """
         return self._list(\
             self._org_api_v2_base_url, \