+++ terrasnek/runs.py
@@ -32,8 +32,9 @@ class TFCRuns(TFCEndpoint):
         """
         ``GET /workspaces/:workspace_id/runs``
 
-        This endpoint supports pagination with standard URL query parameters; remember to
-        percent-encode.
+        Query parameter(s) (`details <https://www.terraform.io/docs/cloud/api/run.html#query-parameters>`_):
+            - ``page`` (Optional)
+            - ``page_size`` (Optional)
         """
         url = f"{self._ws_api_v2_base_url}/{workspace_id}/runs"
         return self._list(url, page=page, page_size=page_size)