+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.8] - TBD
+## [0.0.9] - TBD
 
 - TBD
 
-## [0.0.7] - TBD
+## [0.0.8] - 2020-09-17
+
+- Adding more information on query parameters for each relevant endpoint.
+- Fix a bug in the Team Access API.
+- Update the documentation to add the latest endpoints supported.
+
+## [0.0.7] - 2020-08-26
 
 - Added Agents API for Terraform Cloud Business Tier
 - Added Agents Tokens API for Terraform Cloud Business Tier